debugging, cleaning and refactoring code
