+++ projects/client/src/hoc/authProvider.js
