+++ projects/client/src/pages/WarehouseMapPage.js
