+++ projects/client/src/pages/admin/AdminManageDataPage.js
