+++ projects/server/src/service/stockHistoryService.js
